+++ utils/modelUtils.py
@@ -35,8 +35,8 @@ def get_conv_output_widths(input_widths: torch.Tensor, conv_parameters: dict = N
         kernel_size, stride, padding = conv_cfg['kernel_size'], conv_cfg['stride'], conv_cfg['padding']
         input_widths = torch.floor(((input_widths + 2 * padding[1] - (kernel_size[1] - 1) - 1) / stride[1]) + 1)
 
-    return input_widths.type(torch.int64)
+    return input_widths.type(torch.int32)
 
 
 def token_mask_to_lengths(token_mask: torch.Tensor) -> torch.Tensor:
-    return (1 - token_mask).sum(-1).type(torch.int)
+    return (1 - token_mask).sum(-1).type(torch.int32)
